+++ src/compiler/special_attrs.ml
@@ -12,7 +12,7 @@ let svg = "svg"
 let pdf_resolution = "pdf-resolution"
 let pause_block = "pause-block"
 let as_html = "as-html"
-let auto_next = "auto-next"
+let auto_continue = "auto-continue"
 
 let all_attrs =
   [
@@ -30,5 +30,5 @@ let all_attrs =
     pdf_resolution;
     pause_block;
     as_html;
-    auto_next
+    auto_continue
   ]