+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Added
 
-- `auto-next` attribute to automatically continue to the next action (#240)
+- `auto-continue` attribute to automatically continue to the next action (#240)
 
 ## [v0.11.0] Brazlip (Sunday, the 24th of May, 2026)